% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fBD2.R
\name{fBD2}
\alias{fBD2}
\title{second type of fast modified band depth calculation for fda
 
Method for fast modified band depth calculation. Different from fMBD. Find out what motivated this difference.}
\usage{
fBD2(data)
}
\arguments{
\item{data}{name of dataset}
}
\description{
second type of fast modified band depth calculation for fda
 
Method for fast modified band depth calculation. Different from fMBD. Find out what motivated this difference.
}
\author{
Ying Sun and Marc G.Genton
}

