% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBD.R
\name{MBD}
\alias{MBD}
\title{modified band depth calculation for fda
 
Method for modified band depth (fMBD) calculation}
\usage{
MBD(x, xRef = NULL)
}
\arguments{
\item{x}{name of dataset}

\item{xRef}{Edit this}
}
\description{
modified band depth calculation for fda
 
Method for modified band depth (fMBD) calculation
}
\author{
Ying Sun and Marc G.Genton
}

