% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationTest.R
\name{permutationTest}
\alias{permutationTest}
\title{Implements permutation test based on band depth calculation}
\usage{
permutationTest(controlData, testData, n_perm)
}
\arguments{
\item{controlData}{Data for control subjects, each column must be a subject}

\item{testData}{Data for test subjects, each column must be a subject}

\item{n_perm}{number of permutations to run}
}
\description{
This function takes a PCA decomposition of a PET dataset and simulates new images.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

