% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationTestdisp.R
\name{permutationTestdisp}
\alias{permutationTestdisp}
\title{A permutation test for testing dispersion}
\usage{
permutationTestdisp(controlData, testData, n_perm)
}
\arguments{
\item{controlData}{Data for control subjects, each column must be a subject}

\item{testData}{Data for test subjects, each column must be a subject}

\item{n_perm}{number of permutations to run}
}
\description{
This function takes a a control and a test dataset and permforms a permutation test for dispersion
}
\author{
Sara Lopez-Pintado \email{sl2929@columbia.edu}
}

