% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet2D-data.R
\docType{data}
\name{pet2D}
\alias{pet2D}
\title{Simulated 2-dimensional PET data for control subjects.}
\format{A matrix made up of \describe{ \item{"pet2D"}{A 7505 x 50
matrix of simulated PET data. Each column is a vector for one subject which represents a 95 x 79 pixel image. To view the image for the ith
column, use image.i = matrix(pet2D[,i], nrow = 95, ncol = 79) and image(image.i). There are 50 subjects in total.}
}}
\description{
Simulated two-dimensional PET imaging data. The original data is proprietery, but consists of images from
39 control subjects and 29 subjects with major depressive disorder.
The images represent maps of binding potential of 5- hydroxytryptamine (serotonin) 1A receptors (5-HT1A), 
which are thought to play an important role in major depressive disorder. We decomposed the 39 control subjects using 
principal components analysis (PCA). The data provided here are rebuilt from the resulting eigenvalues and eigenvectors of the PCA 
decomposition, along with randomly generated subject-specific scores.
}
\references{
Parsey et al (2006).
Altered Serotonin 1A Binding in Major Depression: A [carbonyl-C-11]WAY100635 Positron Emission Tomography Study
\emph{Biological Psychiatry}, Vol 59 Issue 2, 106 - 113.
}
\keyword{datasets}

