% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processResults.R
\name{processResults}
\alias{processResults}
\title{process Results returned from a simulation}
\usage{
processResults(resultList, nSims = 100, alpha = 0.05)
}
\arguments{
\item{resultList}{Results from a simulation study, invludes pvalues for each simulation step.}

\item{nSims}{number of simulations that were run. Defaults to 100.}

\item{alpha}{Type I error rate.}
}
\description{
This function calculates and returns the rejection rate when given a list of simulation results
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

