% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSumTest.R
\name{rankSumTest}
\alias{rankSumTest}
\title{Implements rank-sum test based on band depth calculation}
\usage{
rankSumTest(controlData, testData, n, m)
}
\arguments{
\item{controlData}{Data for control subjects, each column must be a subject}

\item{testData}{Data for test subjects, each column must be a subject}

\item{n}{Edit this documentation}

\item{m}{Edit this documentation}
}
\description{
This function samples the reference group several times and gets the average p-value of the rank test.
}
\author{
Sara Lopez-Pintado \email{sl2929@columbia.edu}
}

