% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalep.R
\name{scalep}
\alias{scalep}
\title{A permutation test for testing dispersion}
\usage{
scalep(Data, pr)
}
\arguments{
\item{Data}{A matrix of functions where each row is an observed function or image}

\item{pr}{ask Sara}
}
\description{
This function computes an alternative dispersion measure for a sample of functions or images.
The output is a positive scalar that measures the dispersion of the data.
}
\author{
Sara Lopez-Pintado \email{sl2929@columbia.edu}
}

