% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateBrains.R
\name{simulateBrains}
\alias{simulateBrains}
\title{A function for simulating PET data}
\usage{
simulateBrains(PCAobject, npc = NULL, pve = 0.95, nsubjs = 30,
  scoreMean = 0, muScale = 1, muShift = 0, varscale = 1)
}
\arguments{
\item{PCAobject}{The PCA breakdown of either your test or control data}

\item{npc}{Number of principal components to use for simulating data. Defaults to NULL.}

\item{pve}{Percent Variance explained by PCA decomposition. Defaults to 0.95.}

\item{nsubjs}{The number of images you want to simulate.}

\item{scoreMean}{The mean score for the simulated subjects.}

\item{muScale}{A constant multiplier which scales the mean image.}

\item{muShift}{A constant for shifting the mean image.}

\item{varscale}{A constant multiplier for increasing the variance of all eigenvalues}
}
\description{
This function takes a PCA decomposition of a PET dataset and simulates new images.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

