% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumf.R
\name{volumf}
\alias{volumf}
\title{A permutation test for testing dispersion}
\usage{
volumf(Data)
}
\arguments{
\item{Data}{A matrix of functions where each row is an observed function/image}
}
\description{
This function will compute a dispersion measure for a sample of functions/images
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

